requirements.txt
torch
torchvision
Pillow
colorlog
numpy
psutil
scikit-image
kornia
